requirements.txt
starlette==0.14.2
pydantic==1.8.2
fastapi==0.68.1
mysql_connector_repackaged==0.3.1
